feat: Adds basic implementation of moving avatar
After considering some options, I chose to use a Framelayout on which
the avatar, the background and a constraintlayout are placed. The
constraintlayout defines clickable areas by using guidelines. As the
number of clickable areas should remain rather small, this is (imo)
still OK.

The alternatives, working with pre-defined areas defined by coordinates
doesn't seem an option because this doesn't work on all screen
resolutions.