feat: Add event details screen
Adds the screen in which the user enters the details of a new event.
This screen has now also been linked to the main screen (city overview).
From the event details screen the user can also go to the DrawEmotion-
AvatarScreen.

This also required replacing the "description" attribute from an Event
with a "title" attribute, resulting in some database changes