fix: Add function arg to call at end of animation
Previously every call to moveAvatarToLocationOf triggered the same
fragment opening.
This was made more generic by providing an extra argument to that
function which specificies the action to do once the animation is done.
This allows you to open whatever fragment you like, or do another action.
This could be a possible way to animate moving to several points on the
map.

FAITH-27