refactor: replace DetailType enum with subclasses
This was harder than expected because Room doesn't easily support
class hierarchies. The DAO can't return an abstract supertype so every
subclass requires its own query, and the result of all these queries
have to be combined.

This also has an effect on the EventWithDetails relation, which now
contains 3 lists for each type of detail, instead of just one.

This also had an effect on the ests. All changes are contained in the
database layer however.
FAITH-32