refactor/fix: for io thread when building observables in GetUserUseCase
This is an issue that also showed itself in other use cases that did
network calls: when building observables in the use case they will use
the main thread as subscriber by default, even though in the base Use
case the IO scheduler is defined as the subscriber.

The fix is to specifically mention the IO scheduler when building the
observables that are part of the use case. To do this in a clean way,
the subscriber is now also passed as a constructor parameter for all use
cases.
As to not have to redefine all constructors in the Koin modules, the IO
scheduler was defined as a default argument for the new  subscriber
argument.

In tests, the Schedulers.trampoline() will have to be passed explicitly
to ensure synchronous operation.